feat(admin): unify faculty and staff management view
